Initial commit
Initial implementation of a master detail application that presents a list of videos.  Tapping on one of the videos shows a screen with more details about the selected video.  Scrolling to the bottom of the list causes more videos to be fetched.  All dates are presented according to the current locale and all durations are shown to the nearest minute.

The code is architected according to MVVM and the project is arranged accordingly - one group for the model, one for the view model and one for the view.  Unit tests live in the GMBN BrowserTests target.

The view and view model are largely complete.  For now the model uses a video repository that returns test data, rather than data from the real YouTube API.