Add YouTube client
The client offers enough support to:
- Fetch the IDs of all videos for a given channel, via the /search endpoint
- Fetch the title, description, thumbnail, duration and published date / time, via the /videos endpoint

This should provide enough functionality to implement a VideoRepository that can fetch results in ascending date order.

For now no production code uses the client, but it can be tested using the tests in YouTubeIntegrationTests.  Note that to run these tests you'll need to add your API key to ApiKeys.youTube and rename the test prefixes from skip to test.  In a real codebase these tests would live in a separate target.